fixed unit test

+++ Common/Utilities/ByteArray.cpp
@@ -134,7 +134,7 @@ size_t ByteArray::capacity() const
 //-------------------------------------------------------
 size_t ByteArray::capLength(int iLength) const
 {
-    int l = iLength;
+    size_t l = (size_t)iLength;
     if( l < 0 || l > size() )
     { l = size(); }
     return l;
